requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
annoy==1.17.3
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
banks==2.1.1
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
cymem==2.0.11
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
Django==5.0.14
django-mssql-backend==2.8.1
docopt==0.6.2
docx==0.2.4
emoji==2.14.1
faiss-cpu==1.10.0
filelock==3.18.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.3.2
git-filter-repo==2.47.0
greenlet==3.1.1
griffe==1.7.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.49
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langgraph==0.3.20
langgraph-checkpoint==2.0.23
langgraph-prebuilt==0.1.4
langgraph-sdk==0.1.58
langsmith==0.3.19
legacy-cgi==2.6.3
llama-index-core==0.12.31
llama-index-vector-stores-faiss==0.3.0
logging==0.4.9.6
lxml==5.3.1
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
mssql-django==1.5
multidict==6.2.0
murmurhash==1.0.12
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
odfpy==1.4.1
ollama==0.4.8
openai==1.69.0
orjson==3.10.16
ormsgpack==1.9.0
packaging==24.2
pandas==2.2.3
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.1.0
pipreqs==0.4.13
platformdirs==4.3.7
playwright==1.50.0
preshed==3.0.9
propcache==0.3.0
protobuf==6.30.2
py-unused-deps==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyee==12.1.1
PyMuPDF==1.25.4
pyodbc==5.2.0
pypandoc==1.15
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.1
setuptools==80.4.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
soupsieve==2.6
SQLAlchemy==2.0.39
sqlparse==0.5.3
stanza==1.10.1
sympy==1.13.1
tenacity==9.0.0
textblob==0.19.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
yarg==0.1.10
yarl==1.18.3
zstandard==0.23.0